+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * **New Resource:** `pingdirectory_soft_delete_policy ` (#210)
 * **New Resource:** `pingdirectory_token_claim_validation` (#211)
 * **New Resource:** `pingdirectory_uncached_attribute_criteria` (#212)
+* **New Resource:** `pingdirectory_uncached_entry_criteria` (#213)
 
 
 # v0.6.0 May 23 2023